alpha version - basic program, input in code
